use a template approach for QgsSettingsEntry implementation
