+++ src/core/CMakeLists.txt
@@ -800,6 +800,7 @@ set(QGIS_CORE_SRCS
 
   settings/qgssettings.cpp
   settings/qgssettingsentry.cpp
+  settings/qgssettingsentryimpl.cpp
   settings/qgssettingsregistry.cpp
   settings/qgssettingsregistrycore.cpp
 
@@ -1736,6 +1737,8 @@ set(QGIS_CORE_HDRS
 
   settings/qgssettings.h
   settings/qgssettingsentry.h
+  settings/qgssettingsentryenumflag.h
+  settings/qgssettingsentryimpl.h
   settings/qgssettingsregistry.h
   settings/qgssettingsregistrycore.h
